+++ README.md
@@ -4,11 +4,11 @@
 - controller 
     - ❌게임 진행하는 메소드 (newGame)
     - ❌게임 재시작하는 메소드 (restartGame)
-- dealer
+- computer
     - ✅게임 진행 확인용 boolean 필드 (isGameGoing)
     - ✅서로 다른 3자리 수가 담길 List 필드 (threeDifferentNumbers)
     - ✅서로 다른 3자리 수를 생성하고 불러오는 메소드 (getNewThreeDifferentNumbers)
-- player
+- user
     - ✅서로 다른 3자리 수가 담길 List 필드 (threeDifferentNumbers)
     - ✅서로 다른 3자리 수를 불러오는 메소드 (createThreeDifferentNumbers)
         - ✅입력값의 유효성을 검사하는 메소드 (checkInput)
