+++ src/test/java/mytest/ResultViewTest.java
@@ -2,7 +2,7 @@ package mytest;
 
 import static org.assertj.core.api.Assertions.*;
 
-import baseball.ResultViewer;
+import baseball.settings.ResultView;
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
 import java.io.PrintStream;
@@ -10,14 +10,14 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class ResultViewerTest {
+public class ResultViewTest {
     @DisplayName("볼과 스트라이크 개수가 예상한 프린트문과 똑같이 출력되는지 확인")
     @Test
     void case1() {
         List<Integer> scores = List.of(0, 3);
         OutputStream actualPrintout = new ByteArrayOutputStream();
         System.setOut(new PrintStream(actualPrintout));
-        ResultViewer.printOut(scores);
+        ResultView.printOut(scores);
 
         OutputStream expectedPrintout = new ByteArrayOutputStream();
         System.setOut(new PrintStream(expectedPrintout));
@@ -31,7 +31,7 @@ public class ResultViewerTest {
         List<Integer> scores = List.of(2, 0);
         OutputStream actualPrintout = new ByteArrayOutputStream();
         System.setOut(new PrintStream(actualPrintout));
-        ResultViewer.printOut(scores);
+        ResultView.printOut(scores);
 
         OutputStream expectedPrintout = new ByteArrayOutputStream();
         System.setOut(new PrintStream(expectedPrintout));
@@ -45,7 +45,7 @@ public class ResultViewerTest {
         List<Integer> scores = List.of(0, 0);
         OutputStream actualPrintout = new ByteArrayOutputStream();
         System.setOut(new PrintStream(actualPrintout));
-        ResultViewer.printOut(scores);
+        ResultView.printOut(scores);
 
         OutputStream expectedPrintout = new ByteArrayOutputStream();
         System.setOut(new PrintStream(expectedPrintout));
@@ -59,7 +59,7 @@ public class ResultViewerTest {
         List<Integer> scores = List.of(1, 1);
         OutputStream actualPrintout = new ByteArrayOutputStream();
         System.setOut(new PrintStream(actualPrintout));
-        ResultViewer.printOut(scores);
+        ResultView.printOut(scores);
 
         OutputStream expectedPrintout = new ByteArrayOutputStream();
         System.setOut(new PrintStream(expectedPrintout));
